build: ditch dynamic vcs versioning
Users don't really see or care about this, it was only useful for devs
running local copies. Implied in this change is forgoing any CI that
produces build artifacts on every commit. Cool, but not necessary or
likely to be used on a project of this scale.

Not requiring a git repository (and history) at build time simplifies
other CI, and means a wheel can actually be built from sdist
(important).

The pyproject.toml will be the single-source of truth for the package
version. The release-please-manifest needs to stay, but hopefully will
be kept in sync by the Action