feat: enable color argparse help on python >= 3.14
Cool new feature we can take advantage of

Also enabled argument correction suggests on python >= 3.14.
This won't have any effects with the current arguments, but added for
posterity.

fix: absolute path to script in help output on python >= 3.14

This is due to a change to the default `prog` attribute in argparse. It
was meant to prevent '__main__' appearing in the program output when
invoked as a module. But, it breaks all programs that take advantage of
installable entrypoints....

Not sure what they were intending for this. In our case, I override the
name since we only support running from the entrypoint